Close base layers control in custom manner
Close base layers control in custom manner. Do so on baselayerchange.
For mobile devices, default behavior is to close it with a map click but
this app initiates a new camp search with a map click so there is a
conflict (i.e. - I don't want a new camp search when closing the
control). Note that the problem still exists if user doesn't select a
new basemap; they have to click on the map to close base layers control
so I should come up with a more complete fix at some point.